fix: Visitor always uses the correct CIDs
We used `identityHasherChooser` here on the theory that the hash would
be thrown away, since we're generating it again later, and it's better
not to do it twice for no reason. Unfortunately, that's not quite right:
the hash from the hasher chosen by the `HasherChooser` is used for the
link returned as the root node link.