refactor: Remove `DAGScanStateAwaitingChildren`
We don't actually need it. It's almost exactly the same thing as the
pending state for a directory. The only time it makes a difference is if
the process is interrupted after the directory notices the children are
completed but before completing itself, and the consequence is an
additional check in the DB. The check is also cheaper now, as it's a
single query.