feat: After an error, can be retried safely (#112)
Big changes here!

* Adjusted the workers/APIs/domains to be much more consistent.
* Removed most of the upload states. An upload is now either pending or
started. If an upload is started, it's failed if it has an error
message, and it's successful if it has a `rootCID`. The state of the
file scans, DAG scans, and shards are in those models.

Now, if the process fails at any point, the data remains frozen in the
intermediate state it failed in, and can be restarted from there.

Notably, this does *not* directly address the underlying data on disk
changing during the process, but that should also be easier to deal with
now, as we can simply delete facts from the DB that we know are no
longer true and idempotently throw the whole process at it again. But
doing that is still for a future PR.








#### PR Dependency Tree


* **PR #112** 👈
  * **PR #115**
    * **PR #116**
      * **PR #117**
        * **PR #118**
          * **PR #124**
            * **PR #125**
              * **PR #126**
                * **PR #127**
                  * **PR #128**

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)