feat: Blobs are automatically replicated at 3 (#135)
Closes #119 

Notably, while the original issue was titled "Replicate at end of guppy
upload", I've opted to replicate at the end of each blob add and not all
at once at the end. I think that's the right call? Happy to talk it out
if there's disagreement, though.







#### PR Dependency Tree

* **PR #134**
  * **PR #135**  👈
    * **PR #136**
      * **PR #137**
        * **PR #138**
          * **PR #139**
            * **PR #140**

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)