fix: Don't add links for directory children from other spaces (#229)
This was nasty to track down.

If you reuse the same DB and upload the same directory multiple times,
you'll end up adding the directory children from _previous_ runs to your
current UnixFS directory. Of course, UnixFS is fine with this; it's not
until you try to _retrieve_ the data and write it to disk that you have
a problem, because you're trying to write the same file twice. UnixFS
directories can contain the same file multiple times, but not real ones.

The solution is to add a missing space check. As soon as we descend into
a directory and look up its children, we're picking up the children from
previous runs, and generating links from those. This fixes that.