fix: `AddShardsForUpload` is correctly idempotent (#240)
This means that if `space/blob/replicate` fails (or `filecoin/offer`,
though we haven't seen that happen so far), we won't `space/blob/add`
again on retry. Originally, `space/blob/add` was idempotent, but on
Forge, it's only "idempotent" if it happens to land on the same SP;
otherwise it will add it a second time, to the new SP. With this change,
we hold onto the location at the original SP and can restart from there
on retry.




#### PR Dependency Tree


* **PR #240** 👈

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)